Refactored multi-sim tests with new MultSimTestHarness generic class
