+++ tests/test_multisim_materials_cmfd/test_multisim_materials_cmfd.py
@@ -4,21 +4,18 @@ import os
 import sys
 sys.path.insert(0, os.pardir)
 sys.path.insert(0, os.path.join(os.pardir, 'openmoc'))
-from testing_harness import TestHarness
+from testing_harness import MultiSimTestHarness
 from input_set import GridInput
 import openmoc
 
 
-class MultiSimMaterialsCmfdTestHarness(TestHarness):
+class MultiSimMaterialsCmfdTestHarness(MultiSimTestHarness):
     """A multi-simulation eigenvalue calculation with different materials in
      a grid geometry with CMFD."""
 
     def __init__(self):
         super(MultiSimMaterialsCmfdTestHarness, self).__init__()
         self.input_set = GridInput()
-        self.num_simulations = 3
-        self.num_iters = []
-        self.keffs = []
 
     def _create_geometry(self):
         """Initialize CMFD and add it to the Geometry."""
@@ -55,21 +52,8 @@ class MultiSimMaterialsCmfdTestHarness(TestHarness):
                 materials[material_id].thisown = 1
 
             # Run eigenvalue calculation and store the results
+            self.num_simulations = 1
             super(MultiSimMaterialsCmfdTestHarness, self)._run_openmoc()
-            self.num_iters.append(self.solver.getNumIterations())
-            self.keffs.append(self.solver.getKeff())
-
-    def _get_results(self, num_iterations=True, keff=True, fluxes=False,
-                     num_fsrs=False, num_tracks=False, num_segments=False,
-                     hash_output=False):
-        """Return eigenvalues from each simulation into a string."""
-
-        # Write out the iteration count and eigenvalues from each simulation
-        outstr = ''
-        for num_iters, keff in zip(self.num_iters, self.keffs):
-            outstr += 'Iters: {0}\tkeff: {1:12.5E}\n'.format(num_iters, keff)
-
-        return outstr
 
 
 if __name__ == '__main__':
