+++ tests/test_multisim_num_groups/test_multisim_num_groups.py
@@ -7,20 +7,19 @@ import numpy as np
 
 sys.path.insert(0, os.pardir)
 sys.path.insert(0, os.path.join(os.pardir, 'openmoc'))
-from testing_harness import TestHarness
+from testing_harness import MultiSimTestHarness
 from input_set import HomInfMedInput
 import openmoc
 
 
-class MultiSimNumGroupsTestHarness(TestHarness):
+class MultiSimNumGroupsTestHarness(MultiSimTestHarness):
     """A multi-simulation eigenvalue calculation for a homogeneous infinite
     medium with 1-group and 2-group data."""
 
     def __init__(self):
         super(MultiSimNumGroupsTestHarness, self).__init__()
         self.input_set = HomInfMedInput()
-        self.num_iters = []
-        self.keffs = []
+        self.num_simulations = 1
 
     def _run_openmoc(self):
         """Run multiple OpenMOC eigenvalue calculations with 1- and 2-group
@@ -39,8 +38,6 @@ class MultiSimNumGroupsTestHarness(TestHarness):
 
         # Run eigenvalue calculation and store the results
         super(MultiSimNumGroupsTestHarness, self)._run_openmoc()
-        self.num_iters.append(self.solver.getNumIterations())
-        self.keffs.append(self.solver.getKeff())
 
         # Setup 2-group multi-group cross sections
         material.setName('2-group infinite medium')
@@ -52,20 +49,6 @@ class MultiSimNumGroupsTestHarness(TestHarness):
 
         # Run eigenvalue calculation and store the results
         super(MultiSimNumGroupsTestHarness, self)._run_openmoc()
-        self.num_iters.append(self.solver.getNumIterations())
-        self.keffs.append(self.solver.getKeff())
-
-    def _get_results(self, num_iterations=True, keff=True, fluxes=False,
-                     num_fsrs=False, num_tracks=False, num_segments=False,
-                     hash_output=False):
-        """Return eigenvalues from each simulation into a string."""
-
-        # Write out the iteration count and eigenvalues from each simulation
-        outstr = ''
-        for num_iters, keff in zip(self.num_iters, self.keffs):
-            outstr += 'Iters: {0}\tkeff: {1:12.5E}\n'.format(num_iters, keff)
-
-        return outstr
 
 
 if __name__ == '__main__':
